This fixes #2
Updating the script to account for possible relative path passed to the script as the launch file argument.

The script relies on the absolute path in order to determine the package name that contains the launch file. Relative paths that do not include the package folder (i.e., folder directly before the 'launch' folder) previously caused an error when they should not. This has been resolved.