This fixes #3
The issue is that dot nodes pertaining to launch files, ROS nodes, and rosparam files were not given unique names such that two files (or nodes) that share a name but are in different packages were represented by the same dot node in the graph. This should not have been the case.

The code has been updated to assign a unique name for all launch files, ROS nodes, and yaml files so that the containing package is included in the name for the dot node. This allows them to be properly identified in the graph.