This fixes #5
An error was being specifically raised in the case where an arg did not specify either a "default" value or a "value". ROS does not enforce this condition, so it was invalid to raise an error in this case.

This change eliminates that error so that an error will only be generated when an undefined argument gets generated, thus using:

    <arg name='name' />

Should no longer generate an error.