This fixes #4
Updated the script to properly check for cycles in the graph by looking at the ancestor tree of a launch file to determine if it already exists in the path to itself.

Previously it was erroneously identifiying any launch file that was included more than once as a cycle, which it is not.