This fixes #1 (again)
Bug was introduced breaking env due to a typo, and attempt to get the package name for rosparams was broken (it was using the un-substituted path instead of the resolved path which meant it could never find a true package for the rosparam). Both of these problems have been solved.