Created several unit tests for validating the behavior of the roslaunch-to-dot script.
The unit tests are stored under the test directory and should be run by doing:

    $ cd roslaunch_to_dot
    $ ./run_tests.sh

Updated the script to draw missing rosparam files as filled red nodes to be consistent.

Updated the error message displayed when the script is given a launch file with an invalid type. The valid filetypes are present in the error message now.